% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarge_plot_custom.R
\name{sarge_plot_custom}
\alias{sarge_plot_custom}
\title{Plot sarge data using custom variables}
\usage{
sarge_plot_custom(data, start, end, vars, colors)
}
\arguments{
\item{data}{data to plot}

\item{start}{start date}

\item{end}{end date}

\item{vars}{variables to plot}

\item{colors}{colors to use}
}
\value{
ggplot of sarge data
}
\description{
Plot sarge data using custom variables
}
