% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarge_plot_defaults.R
\name{sarge_plot_defaults}
\alias{sarge_plot_defaults}
\title{Plot SARGE data using default variables}
\usage{
sarge_plot_defaults(data, data_type, start, end, variables)
}
\arguments{
\item{data}{full sarge dataset, stored as a list containing multiple datastreams}

\item{data_type}{data type to plot (e.g., "GHG")}

\item{start}{start date for plot}

\item{end}{end date for plot}

\item{variables}{dataframe of variables and colors}
}
\value{
Plot of SARGE data
}
\description{
Plot SARGE data using default variables
}
