% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarge_read.R
\name{sarge_read}
\alias{sarge_read}
\title{Read in SARGE data}
\usage{
sarge_read(files, variables, timezone, resolution = "minute")
}
\arguments{
\item{files}{dataframe with files to load}

\item{variables}{dataframe with variable descriptions}

\item{timezone}{timezone}

\item{resolution}{temporal resolution (e.g., "minute" or "day")}
}
\value{
list with multiple sarge dataframes
}
\description{
Read in SARGE data
}
